+++ Engine/Source/Tempest/cirect.h
@@ -50,28 +50,28 @@ namespace NTempest {
     ~CiRect() {
     }
 
-    void Get(long &top, long &left, long &bottom, long &right) const;
-    void Set(long top, long left, long bottom, long right);
-    operator tagRECT() const;
+    void    Get(long &top, long &left, long &bottom, long &right) const;
+    void    Set(long top, long left, long bottom, long right);
+            operator tagRECT() const;
     CiRect &operator+=(const CiRect &value);
     CiRect &operator-=(const CiRect &value);
     CiRect &operator*=(const CiRect &value);
     CiRect &operator/=(const CiRect &value);
-    CiRect operator-() const;
-    void Stretch(const C2iVector &value);
-    void Stretch(long horizontal, long vertical);
-    void Offset(const C2iVector &value);
-    void Offset(long horizontal, long vertical);
-    bool NotEmpty() const;
-    bool Empty() const;
-    bool Invalid() const;
-    bool NotInvalid() const;
-    bool Encloses(const CiRect &value) const;
-    bool Encloses(const C2iVector &value) const;
-    bool Contains(const CiRect &value) const;
-    bool Contains(const C2iVector &value) const;
-    bool InOpenR(const CiRect &value) const;
-    bool InOpenR(const C2iVector &value) const;
+    CiRect  operator-() const;
+    void    Stretch(const C2iVector &value);
+    void    Stretch(long horizontal, long vertical);
+    void    Offset(const C2iVector &value);
+    void    Offset(long horizontal, long vertical);
+    bool    NotEmpty() const;
+    bool    Empty() const;
+    bool    Invalid() const;
+    bool    NotInvalid() const;
+    bool    Encloses(const CiRect &value) const;
+    bool    Encloses(const C2iVector &value) const;
+    bool    Contains(const CiRect &value) const;
+    bool    Contains(const C2iVector &value) const;
+    bool    InOpenR(const CiRect &value) const;
+    bool    InOpenR(const C2iVector &value) const;
 
     long Width() const {
       return r - l;
@@ -81,28 +81,28 @@ namespace NTempest {
       return b - t;
     }
 
-    void SetWidth(long value);
-    void SetHeight(long value);
-    C2iVector TopLeft() const;
-    C2iVector TopRight() const;
-    C2iVector BottomLeft() const;
-    C2iVector BottomRight() const;
-    void Center(const CiRect &value);
-    C2iVector Center() const;
-    C2iVector Diagonal() const;
-    void CenterV(const CiRect &value);
-    void CenterH(const CiRect &value);
-    void AlignTop(const CiRect &value);
-    void AlignLeft(const CiRect &value);
-    void AlignBottom(const CiRect &value);
-    void AlignRight(const CiRect &value);
+    void          SetWidth(long value);
+    void          SetHeight(long value);
+    C2iVector     TopLeft() const;
+    C2iVector     TopRight() const;
+    C2iVector     BottomLeft() const;
+    C2iVector     BottomRight() const;
+    void          Center(const CiRect &value);
+    C2iVector     Center() const;
+    C2iVector     Diagonal() const;
+    void          CenterV(const CiRect &value);
+    void          CenterH(const CiRect &value);
+    void          AlignTop(const CiRect &value);
+    void          AlignLeft(const CiRect &value);
+    void          AlignBottom(const CiRect &value);
+    void          AlignRight(const CiRect &value);
     static CiRect Intersection(const CiRect &a, const CiRect &b, const CiRect &clip);
     static CiRect Intersection(const CiRect &left, const CiRect &right);
     static CiRect Union(const CiRect &left, const CiRect &right);
     static CiRect ClippedLocal(const CiRect &value, const CiRect &clip);
-    static unsigned long Difference(const CiRect &left, const CiRect &right, CiRect *result);
-    CiRect Intersect(const CiRect &right);
-    CiRect Unite(const CiRect &right);
+    static DWORD  Difference(const CiRect &left, const CiRect &right, CiRect *result);
+    CiRect        Intersect(const CiRect &right);
+    CiRect        Unite(const CiRect &right);
   };
 
 }  // namespace NTempest
