+++ Engine/Source/Tempest/crect.h
@@ -60,19 +60,19 @@ namespace NTempest {
       r = right;
     }
 
-    void Get(float &top, float &left, float &bottom, float &right) const;
-    operator tagRECT() const;
-    CRect asCRect() const;
+    void         Get(float &top, float &left, float &bottom, float &right) const;
+                 operator tagRECT() const;
+    CRect        asCRect() const;
     const CRect *asFloatPtr() const;
-    CRect &operator+=(const CRect &value);
-    CRect &operator-=(const CRect &value);
-    CRect &operator*=(const CRect &value);
-    CRect &operator/=(const CRect &value);
-    CRect operator-() const;
-    void Stretch(const C2Vector &value);
-    void Stretch(float horizontal, float vertical);
-    void Offset(const C2Vector &value);
-    void Offset(float horizontal, float vertical);
+    CRect       &operator+=(const CRect &value);
+    CRect       &operator-=(const CRect &value);
+    CRect       &operator*=(const CRect &value);
+    CRect       &operator/=(const CRect &value);
+    CRect        operator-() const;
+    void         Stretch(const C2Vector &value);
+    void         Stretch(float horizontal, float vertical);
+    void         Offset(const C2Vector &value);
+    void         Offset(float horizontal, float vertical);
 
     bool NotEmpty() const {
       return t < b && l < r;
@@ -96,21 +96,21 @@ namespace NTempest {
       return b - t;
     }
 
-    void SetWidth(float value);
-    void SetHeight(float value);
+    void     SetWidth(float value);
+    void     SetHeight(float value);
     C2Vector TopLeft() const;
     C2Vector TopRight() const;
     C2Vector BottomLeft() const;
     C2Vector BottomRight() const;
-    void Center(const CRect &value);
+    void     Center(const CRect &value);
     C2Vector Center() const;
     C2Vector Diagonal() const;
-    void CenterV(const CRect &value);
-    void CenterH(const CRect &value);
-    void AlignTop(const CRect &value);
-    void AlignLeft(const CRect &value);
-    void AlignBottom(const CRect &value);
-    void AlignRight(const CRect &value);
+    void     CenterV(const CRect &value);
+    void     CenterH(const CRect &value);
+    void     AlignTop(const CRect &value);
+    void     AlignLeft(const CRect &value);
+    void     AlignBottom(const CRect &value);
+    void     AlignRight(const CRect &value);
 
     static CRect Lerp(const CRect &a, const CRect &b, const CRect &t);
     static CRect Intersection(const CRect &a, const CRect &b, const CRect &clip);
@@ -126,7 +126,7 @@ namespace NTempest {
 
     static CRect Union(const CRect &left, const CRect &right);
     static CRect ClippedLocal(const CRect &value, const CRect &clip);
-    static unsigned long Difference(const CRect &left, const CRect &right, CRect *result);
+    static DWORD Difference(const CRect &left, const CRect &right, CRect *result);
 
     CRect Intersect(const CRect &right) {
       return Intersection(*this, right);
