+++ WoW/Source/WorldClient/MapClip.cpp
@@ -4,17 +4,17 @@
 
 #include "WorldClient/World.h"
 
-static const unsigned long vCnt[4] = {2, 3, 5, 9};
-static const unsigned long vStp[4] = {8, 4, 2, 1};
+static const DWORD vCnt[4] = {2, 3, 5, 9};
+static const DWORD vStp[4] = {8, 4, 2, 1};
 
 void CMapChunk::UpdateClipBuffer() {
-  int           indexList[9];
-  unsigned long cnt = vCnt[lod];
+  int   indexList[9];
+  DWORD cnt = vCnt[lod];
 
-  unsigned long step = vStp[lod];
-  unsigned long start = CWorldScene::camPos.x > CWorldScene::camTarg.x ? 136 : 0;
-  unsigned long index = 0;
-  unsigned long vertex = 0;
+  DWORD step = vStp[lod];
+  DWORD start = CWorldScene::camPos.x > CWorldScene::camTarg.x ? 136 : 0;
+  DWORD index = 0;
+  DWORD vertex = 0;
   while (vertex < 9) {
     indexList[index++] = start + vertex;
     vertex += step;
